change: Move health check controller out of v1 namespace
