+++ module-server/src/main/kotlin/de/twomartens/timetable/support/controller/HealthCheckController.kt
@@ -1,4 +1,4 @@
-package de.twomartens.timetable.support.controller.v1
+package de.twomartens.timetable.support.controller
 
 import de.twomartens.timetable.property.ServiceProperties
 import io.swagger.v3.oas.annotations.Hidden
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
 
 @RestController
-@RequestMapping(value = ["/timetable/v1"])
+@RequestMapping(value = ["/timetable"])
 class HealthCheckController(private val properties: ServiceProperties) {
